# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admission Application API",
        "description": "REST API for the Secondary School Admission Registration System. Manages applicant biodata, next-of-kin information, and health records.",
        "contact": {
            "email": "admin@school.edu.ng"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
...
        "/applicants": {
            "get": {
                "tags": [
                    "Applicants"
                ],
                "summary": "List applicants with optional filtering and pagination",
                "operationId": "d8d3856442f7942c16e9d4c4b8ceb926",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name, phone or application number",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "john"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "under_review",
                                "approved",
                                "rejected"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of applicants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ApplicantResource"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 95
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Applicants"
                ],
                "summary": "Submit a new admission application",
                "description": "Creates an applicant record together with next-of-kin and health record in one request. An application number is auto-generated. Send as `multipart/form-data` when including a passport photo.",
                "operationId": "2623e0093846ec8cf08ff9816fac438e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Application submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application submitted successfully."
                                        },
                                        "applicant": {
                                            "$ref": "#/components/schemas/ApplicantResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/applicants/lookup": {
            "get": {
                "tags": [
                    "Applicants"
                ],
                "summary": "Look up an application by application number (public)",
                "operationId": "5fb008e6c0d3b1fa09b0cd38e197d809",
                "parameters": [
                    {
                        "name": "applicationNumber",
                        "in": "query",
                        "description": "e.g. APP-6A36A61045B86",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicantResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Application not found"
                    },
                    "422": {
                        "description": "applicationNumber query param is required"
                    }
                }
            }
        },
        "/applicants/{id}": {
            "get": {
                "tags": [
                    "Applicants"
                ],
                "summary": "Get a single application",
                "operationId": "03433cd4041ff20e7f79a4906cb6cf00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Applicant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Applicant details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicantResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Applicants"
                ],
                "summary": "Update an existing application",
                "operationId": "275ea706e9f3e3bcdcf6879c2ce2b889",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Applicant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplicantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Application updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application updated successfully."
                                        },
                                        "applicant": {
                                            "$ref": "#/components/schemas/ApplicantResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Applicants"
                ],
                "summary": "Delete an application",
                "operationId": "77b690f6fa07ba9bdb13fa591abdda95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Applicant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Application deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Application deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApplicantResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "application_number": {
                        "type": "string",
                        "example": "APP-6672A1B2C3D4"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Chukwuemeka"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Okafor"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "James",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "example": "Male",
                        "enum": [
                            "Male",
                            "Female"
                        ]
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08012345678"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "chukwuemeka@example.com"
                    },
                    "passport_url": {
                        "type": "string",
                        "example": "http://localhost/storage/passports/1/photo.jpg",
                        "nullable": true
                    },
                    "home_address": {
                        "type": "string",
                        "example": "12 Aba Road, Owerri"
                    },
                    "state": {
                        "type": "string",
                        "example": "Imo"
                    },
                    "local_government": {
                        "type": "string",
                        "example": "Owerri North"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "under_review",
                            "approved",
                            "rejected"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "next_of_kin": {
                        "$ref": "#/components/schemas/NextOfKinInput"
                    },
                    "health_record": {
                        "$ref": "#/components/schemas/HealthRecordInput"
                    }
                },
                "type": "object"
            },
            "ApplicantRequest": {
                "required": [
                    "first_name",
                    "last_name",
                    "gender",
                    "phone_number",
                    "email",
                    "password",
                    "home_address",
                    "state",
                    "local_government",
                    "next_of_kin",
                    "health_record"
                ],
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "Chukwuemeka"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Okafor"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "James",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "example": "Male",
                        "enum": [
                            "Male",
                            "Female"
                        ]
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08012345678"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "chukwuemeka@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "secret1234",
                        "minLength": 8
                    },
                    "passport": {
                        "description": "Optional passport photo (jpg/jpeg/png, max 2 MB). Send as multipart/form-data when included.",
                        "type": "string",
                        "format": "binary"
                    },
                    "home_address": {
                        "type": "string",
                        "example": "12 Aba Road, Owerri"
                    },
                    "state": {
                        "type": "string",
                        "example": "Imo"
                    },
                    "local_government": {
                        "type": "string",
                        "example": "Owerri North"
                    },
                    "next_of_kin": {
                        "$ref": "#/components/schemas/NextOfKinInput"
                    },
                    "health_record": {
                        "$ref": "#/components/schemas/HealthRecordInput"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token returned by POST /api/v1/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}